refactor: allow hiding side panel while allowing artifacts view
