+++ client/src/components/SidePanel/index.ts
@@ -1,2 +1,2 @@
-export { default as SidePanel } from './SidePanel';
+export { default as SidePanelGroup } from './SidePanelGroup';
 export { default as SideNav } from './Nav';
